% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunction.R
\name{myfunction}
\alias{myfunction}
\title{myfunction}
\usage{
myfunction(user)
}
\arguments{
\item{user}{character string with user's name}
}
\value{
Character string greeting the user
}
\description{
A hello world example
}
